chore: ATLAS-51: Moved files to Utils project
