+++ Atlas.MatchingAlgorithm.MatchingDictionary/Services/DataRetrieval/Dpb1TceGroupLookupService.cs
@@ -1,4 +1,4 @@
-﻿using Atlas.MatchingAlgorithm.Common.Services;
+﻿using Atlas.Utils.Hla.Services;
 using Atlas.MatchingAlgorithm.Common.Models;
 using Atlas.MatchingAlgorithm.MatchingDictionary.Caching;
 using Atlas.MatchingAlgorithm.MatchingDictionary.Models.Lookups.Dpb1TceGroupLookup;
