+++ Atlas.MatchingAlgorithm.MatchingDictionary/Services/DataRetrieval/HlaScoringLookupService.cs
@@ -1,4 +1,5 @@
-﻿using Atlas.MatchingAlgorithm.Common.Services;
+﻿using Atlas.Utils.Hla.Services;
+using Atlas.Utils.Hla.Models;
 using Atlas.MatchingAlgorithm.Common.Models;
 using Atlas.MatchingAlgorithm.MatchingDictionary.Caching;
 using Atlas.MatchingAlgorithm.MatchingDictionary.Models.Lookups;
