+++ Atlas.MatchingAlgorithm.MatchingDictionary/Services/DataRetrieval/Lookups/AlleleStringLookup.cs
@@ -1,4 +1,4 @@
-﻿using Atlas.MatchingAlgorithm.Common.Services;
+﻿using Atlas.Utils.Hla.Services;
 using Atlas.MatchingAlgorithm.Common.Models;
 using Atlas.MatchingAlgorithm.MatchingDictionary.Repositories;
 using System.Collections.Generic;
