+++ Atlas.MatchingAlgorithm/Services/MatchingDictionary/NmdpCodeCachingService.cs
@@ -1,7 +1,8 @@
 ﻿using LazyCache;
 using Nova.HLAService.Client;
 using Atlas.MatchingAlgorithm.Common.Models;
-using Atlas.MatchingAlgorithm.Common.Services;
+using Atlas.Utils.Hla.Services;
+using Atlas.Utils.Hla.Models;
 using Atlas.MatchingAlgorithm.MatchingDictionary.Caching;
 using Nova.Utils.ApplicationInsights;
 using Nova.Utils.Models;
