+++ Atlas.Utils/Atlas.Utils.csproj
@@ -4,4 +4,8 @@
     <TargetFramework>netcoreapp2.2</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Nova.Utils" Version="14.3.0" />
+  </ItemGroup>
+
 </Project>
