+++ Atlas.Utils/Hla/Services/AlleleStringSplitters/AlleleStringOfNamesSplitter.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using Atlas.MatchingAlgorithm.Common.Models.HlaTypings;
+using Atlas.Utils.Hla.Models.HlaTypings;
 
-namespace Atlas.MatchingAlgorithm.Common.Services.AlleleStringSplitters
+namespace Atlas.Utils.Hla.Services.AlleleStringSplitters
 {
     internal class AlleleStringOfNamesSplitter : AlleleStringSplitterBase
     {
